allow encryptfn to take an arbitrary public key
